chore: restrict billing utils to billing
